backend/requirements.txt
# Core web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Authentication
bcrypt==4.0.1
PyJWT==2.8.0

# Templating and internals
Jinja2==3.1.2
Werkzeug==2.3.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7

# Chatbot / AI integration
openai==1.30.1  # Only if you're using OpenAI API

# HTTP requests (for API calls)
requests==2.31.0

# Database (choose one)
SQLAlchemy==2.0.30     # for relational DB (SQLite, PostgreSQL, etc.)
# pymongo==4.6.3       # uncomment if using MongoDB instead

# Environment variable loader
python-dotenv==1.0.1
